project/AgenticRag/backend/app/core/config.py
import os
from dotenv import load_dotenv
from langchain_deepseek import ChatDeepSeek
from langchain_openai import OpenAIEmbeddings

load_dotenv(override=True)

# 路径配置
BASE_DIR = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
UPLOAD_DIR = os.path.join(BASE_DIR, "data", "uploads")
VECTOR_STORE_DIR = os.path.join(BASE_DIR, "data", "vector_stores")

os.makedirs(UPLOAD_DIR, exist_ok=True)
os.makedirs(VECTOR_STORE_DIR, exist_ok=True)

# 模型初始化
def get_llm():
    # 使用 DeepSeek V3 或 V2.5
    return ChatDeepSeek(
        model="deepseek-chat",
        temperature=0.1, # RAG 场景温度低一点更准确
        api_key=os.getenv("DEEPSEEK_API_KEY"),
        base_url=os.getenv("DEEPSEEK_BASE_URL")
    )

def get_embeddings():
    return OpenAIEmbeddings(
        model="text-embedding-3-small",
        api_key=os.getenv("OPENAI_API_KEY"),
        base_url=os.getenv("OPENAI_BASE_URL")
    )