project/AgenticRag/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
langchain-openai
langchain-deepseek
langchain-text-splitters
faiss-cpu
pydantic